Fixed a bug with not differentiable functions
The NOT_DIFFERENTIABLE flag, float("nan"), would not be accepted as an
uncertainty, which makes the whole concept of working with not differentiable
functions broken. Should work now.
Also added a comment regarding this.

Second change are some abbrevations: The arguments to
UncertainVariable.__init__() are no longer called statistic_uncertainty and
systematic_uncertainty, but stat and sys. If they are given as keyword
arguments, the code will be much easier to read.

Also added UVar as an abbrevation to UncertainVariable